Rune system now unrelated to screen resolution

+++ Assets/Scripts/ScriptableObjects/RuneDrawParameters.cs
@@ -6,6 +6,5 @@ public class RuneDrawParameters : ScriptableObject
     [Header("changing this properties doesn't affects already created runes\nPlease recreate them to apply changes")]
     public float distanceBetweenPoints;
     public float acceptableError;
-    public int heavyCheckStep;
+    public float heavyCheckStep;
 }
-
